fix: rename utils
